+++ monitors/c/src/engines/booleanizer.c
